rip GraphicsPreview out of ModelsLibraryDialog
